--- core/src/main/scala-2.x/spire/syntax/CforSyntax.scala
@@ -1,20 +1,0 @@
-package spire
-package syntax
-
-import spire.NoImplicit
-import spire.algebra._
-import spire.algebra.lattice._
-import spire.algebra.partial._
-import spire.math._
-import spire.macros.Syntax
-import spire.syntax.std._
-import scala.annotation.nowarn
-
-trait CforSyntax {
-  def cfor[A](init: A)(test: A => Boolean, next: A => A)(body: A => Unit): Unit =
-    macro Syntax.cforMacro[A]
-  def cforRange(r: Range)(body: Int => Unit): Unit =
-    macro Syntax.cforRangeMacro
-  def cforRange2(r1: Range, r2: Range)(body: (Int, Int) => Unit): Unit =
-    macro Syntax.cforRange2Macro
-}
