+++ core/src/main/scala-2.x/spire/syntax/Syntax.scala
@@ -261,6 +261,15 @@ trait ConvertableFromSyntax {
   implicit def convertableOps[A: ConvertableFrom](a: A): ConvertableFromOps[A] = new ConvertableFromOps(a)
 }
 
+trait CforSyntax {
+  def cfor[A](init: A)(test: A => Boolean, next: A => A)(body: A => Unit): Unit =
+    macro Syntax.cforMacro[A]
+  def cforRange(r: Range)(body: Int => Unit): Unit =
+    macro Syntax.cforRangeMacro
+  def cforRange2(r1: Range, r2: Range)(body: (Int, Int) => Unit): Unit =
+    macro Syntax.cforRange2Macro
+}
+
 trait LiteralsSyntax {
   implicit def literals(s: StringContext): Literals = new Literals(s)
